Split MockClient into MockClient and MockLimiter - as an emulator - identical to the one in the mock-envoy-limiter project
